Create README.md
An in depth phylogenetic analysis was conducted on the 3,367 unique SHAe (IRF3 and p65) to study their evolutionary history across the 100 Vertebrate genomes incorporated in UCSC Browser. An in-house Python-designed bot was used to run automated tasks online, thus facilitating the efficient screening of approximately 3,367 SHAe (IRF3 and p65) for sequence conservation, by applying Multiz 100-way Vertebrates Alignment, phyloP basewise score, and RepeatMasker as the selected tracks of the UCSC Genome Browser. The bot begins by accessing the UCSC Genome Browser site for hg19 genome assembly. Next, rigorous manual analysis was conducted to classify the SHAe according to their evolutionary origins as (1) enriched in Primates (e-Primates), (2) enriched in Mammals (e-Mammals), (3) evolutionarily traced beyond Mammals (b-Mammals). Stringent criteria were applied, including (a) the average phyloP score for each SHAe, (b) sequence similarity between the different genomes more than 60%, (c) the length of the conserved sequences more than 100 bp, and (d) conservation across multiple evolutionary relative organisms. We term these elements “enriched” instead of “specific” to emphasize their distribution in this exact spectrum of 100 vertebrate genomes and to pose the probability of being parts of genomes of organisms that belong to other classes, outside of this cohort, or/and have not yet been entirely sequenced. Hence, we present the snapshots of the genomic coordinates of the 3,367 unique SHAe (IRF3 and p65) as displayed in the UCSC Genome Browser, upon our analyses. Non-repetitive- and repetitive-SHAe surveys are individually charted.